third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferToGPUTexture-texture-readback.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-promise-test.js"></script>
<script src="./webgpu-helpers.js"></script>

<h1>Canvas2D WebGPU Transfer: transferToGPUTexture-texture-readback</h1>
<p>transferToGPUTexture() texture retains the contents of the canvas, and
readback works.</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * transferToGPUTexture() texture should retain the contents of the canvas, and
 * readback works.
 */
canvasPromiseTest(async (canvas) => {
  const device = await getWebGPUDevice();

  // Fill the canvas with a color containing distinct values in each channel.
  const ctx = canvas.getContext('2d');
  ctx.fillStyle = "#4080C0";
  ctx.fillRect(0, 0, 50, 50);

  // Convert the canvas to a texture.
  const tex = ctx.transferToGPUTexture({device: device,
                                        usage: GPUTextureUsage.COPY_SRC});

  // Copy the top-left pixel from the texture into a buffer.
  const buf = copyOnePixelFromTextureAndSubmit(device, tex);

  // Map the buffer and read it back.
  await buf.mapAsync(GPUMapMode.READ);
  const data = new Uint8Array(buf.getMappedRange());

  if (tex.format == 'rgba8unorm') {
    assert_array_equals(data, [64, 128, 192, 255]);
  } else {
    assert_equals(tex.format, 'bgra8unorm');
    assert_array_equals(data, [192, 128, 64, 255]);
  }
}, 'transferToGPUTexture() texture retains the contents of the canvas, ' +
   'and readback works.');

</script>
